MLProject/requirements.txt
joblib==1.4.2
mlflow==2.22.0
mlflow_skinny==2.22.0
pandas==2.2.3
python-dotenv==1.1.0
scikit_learn==1.6.1
